feat: Extract ParticipantTestimonials component for reusable marquee reviews
